third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-ontrack.html

<!DOCTYPE html>
<html>
<head>
<title>RTCPeerConnection-ontrack</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
promise_test(function() {
  let pc = new RTCPeerConnection();
  let eventSequence = [];
  return createStreams({audio:true, video:false}, 1)
    .then(function(streams) {
      pc.addStream(streams[0]);
      let resolvers = Resolver.createResolvers(2);
      pc.ontrack = function() {
        eventSequence.push('ontrack');
        resolvers.shift().resolve();
      }
      setRemoteDescriptionRemoteStreamsMatchLocal(pc).then(function() {
        eventSequence.push('setRemoteDescription');
        resolvers.shift().resolve();
      });
      return Promise.all(Resolver.promises(resolvers));
    }).then(function() {
      // TODO(hbos): We might want to change this so that track events fire
      // before the promise is resolved, see spec discussion
      // https://github.com/w3c/webrtc-pc/issues/1508.
      let expectedEventSequence = [ 'setRemoteDescription',
                                    'ontrack' ];
      assert_array_equals(eventSequence, expectedEventSequence);
    });
}, 'ontrack fires after setRemoteDescription is resolved.');

promise_test(function() {
  let pc = new RTCPeerConnection();
  let eventSequence = [];
  return createStreams({audio:true, video:false}, 1)
    .then(function(streams) {
      pc.addStream(streams[0]);
      let resolvers = Resolver.createResolvers(2);
      pc.onaddstream = function() {
        eventSequence.push('onaddstream');
        resolvers.shift().resolve();
      }
      setRemoteDescriptionRemoteStreamsMatchLocal(pc).then(function() {
        eventSequence.push('setRemoteDescription');
        resolvers.shift().resolve();
      });
      return Promise.all(Resolver.promises(resolvers));
    }).then(function() {
      let expectedEventSequence = [ 'setRemoteDescription',
                                    'onaddstream' ];
      assert_array_equals(eventSequence, expectedEventSequence);
    });
}, 'onaddstream fires after setRemoteDescription is resolved.');

promise_test(function() {
  let pc = new RTCPeerConnection();
  let eventSequence = [];
  return createStreams({audio:true, video:false}, 2)
    .then(function(streams) {
      pc.addStream(streams[0]);
      pc.addStream(streams[1]);
      let resolvers = Resolver.createResolvers(4);
      pc.onaddstream = function(event) {
        eventSequence.push('stream:' + event.stream.id);
        resolvers.shift().resolve();
      };
      pc.ontrack = function(event) {
        eventSequence.push('track:' + event.track.id);
        resolvers.shift().resolve();
      };
      return Promise.all(
          Resolver.promises(resolvers).concat(
              setRemoteDescriptionRemoteStreamsMatchLocal(pc)));
    }).then(function() {
      let expectedEventSequence =
          [ 'stream:' + pc.getRemoteStreams()[0].id,
            'track:' + pc.getRemoteStreams()[0].getAudioTracks()[0].id,
            'stream:' + pc.getRemoteStreams()[1].id,
            'track:' + pc.getRemoteStreams()[1].getAudioTracks()[0].id ];
      assert_array_equals(eventSequence, expectedEventSequence);
    });
}, 'ontrack and onaddstream fires in the expected relative order.');

promise_test(function() {
  let pc = new RTCPeerConnection();
  let eventSequence = [];
  return createStreams({audio:true, video:false}, 1)
    .then(function(streams) {
      pc.addStream(streams[0]);
      let resolvers = Resolver.createResolvers(3);
      let callback = function() {
        if (eventSequence.length == 0) {
          setTimeout(function() {
            eventSequence.push('new task');
            resolvers.shift().resolve();
          }, 0);
        }
        eventSequence.push('track or stream fired');
        resolvers.shift().resolve();
      }
      pc.onaddstream = callback;
      pc.ontrack = callback;
      return Promise.all(
          Resolver.promises(resolvers).concat(
              setRemoteDescriptionRemoteStreamsMatchLocal(pc)));
    }).then(function() {
      assert_array_equals(
          eventSequence,
          [ 'track or stream fired', 'track or stream fired', 'new task' ]);
    });
}, 'ontrack and onaddstream fires in the same task.');

promise_test(function() {
  let pc = new RTCPeerConnection();
  return createStreams({audio:true, video:false}, 1)
    .then(function(streams) {
      pc.addStream(streams[0]);
      let resolver = new Resolver();
      pc.ontrack = function(event) {
        let remoteStream = pc.getRemoteStreams()[0];
        let remoteAudioTrack = remoteStream.getAudioTracks()[0];
        assert_equals(event.track, remoteAudioTrack);
        resolver.resolve();
      };
      return Promise.all([ resolver.promise,
                           setRemoteDescriptionRemoteStreamsMatchLocal(pc) ]);
    });
}, 'ontrack event\'s track matches track in getRemoteStreams().');

promise_test(function() {
  let pc = new RTCPeerConnection();
  return createStreams({audio:true, video:false}, 1)
    .then(function(streams) {
      pc.addStream(streams[0]);
      let resolver = new Resolver();
      pc.ontrack = function(event) {
        let receiver = pc.getReceivers()[0];
        assert_equals(event.receiver, receiver);
        resolver.resolve();
      };
      return Promise.all([ resolver.promise,
                           setRemoteDescriptionRemoteStreamsMatchLocal(pc) ]);
    });
}, 'ontrack event\'s receiver matches getReceivers().');

promise_test(function() {
  let pc = new RTCPeerConnection();
  return createStreams({audio:true, video:false}, 1)
    .then(function(streams) {
      pc.addStream(streams[0]);
      let resolver = new Resolver();
      pc.ontrack = function(event) {
        assert_array_equals(event.streams, pc.getRemoteStreams())
        resolver.resolve();
      };
      return Promise.all([ resolver.promise,
                           setRemoteDescriptionRemoteStreamsMatchLocal(pc) ]);
    });
}, 'ontrack event\'s streams match getRemoteStreams().');

promise_test(function() {
  let pc = new RTCPeerConnection();
  let firedForAudioTrack = false;
  let firedForVideoTrack = false;
  return createStreams({audio:true, video:true}, 1)
    .then(function(streams) {
      pc.addStream(streams[0]);
      let resolvers = Resolver.createResolvers(2);
      pc.ontrack = function(event) {
        let remoteStream = pc.getRemoteStreams()[0];
        if (event.track == remoteStream.getAudioTracks()[0])
          firedForAudioTrack = true;
        else if (event.track == remoteStream.getVideoTracks()[0])
          firedForVideoTrack = true;
        resolvers.shift().resolve();
      };
      return Promise.all(
          Resolver.promises(resolvers).concat(
              setRemoteDescriptionRemoteStreamsMatchLocal(pc)));
    }).then(function() {
      assert_true(firedForAudioTrack && firedForVideoTrack);
    });
}, 'ontrack event fires twice for a stream containing two tracks.');

/**
 * Helper functions.
 */

function createStreams(constraints, numStreams, streamsSoFar = []) {
  if (numStreams == 0) {
    return Promise.resolve(streamsSoFar);
  }
  return navigator.mediaDevices.getUserMedia(constraints)
    .then(function(stream) {
      return createStreams(constraints,
                           numStreams - 1,
                           streamsSoFar.concat([stream]));
    });
}

/**
 * Makes the peer connection's remote streams match its local streams. New local
 * streams result in cloning, removed local streams result in removing the
 * corresponding remote stream.
 *
 * Returns a promise.
 */
function setRemoteDescriptionRemoteStreamsMatchLocal(pc) {
  // In LayoutTests a |MockWebRTCPeerConnectionHandler| is used and as such we
  // cannot establish a real call. Instead we rely on the mock making remote
  // streams match the local streams on |setRemoteDescription|.
  return pc.setRemoteDescription(
      new RTCSessionDescription({type:'answer', sdp:'remote'}));
}

/**
 * The resolver has a |promise| that can be resolved or rejected using |resolve|
 * or |reject|.
 */
class Resolver {
  constructor() {
    let promiseResolve;
    let promiseReject;
    this.promise = new Promise(function(resolve, reject) {
      promiseResolve = resolve;
      promiseReject = reject;
    });
    this.resolve = promiseResolve;
    this.reject = promiseReject;
  }

  static createResolvers(count) {
    let result = [];
    for (let i = 0; i < count; ++i)
      result.push(new Resolver());
    return result;
  }

  static promises(resolvers) {
    let promises = [];
    for (let i = 0; i < resolvers.length; ++i) {
      promises.push(resolvers[i].promise);
    }
    return promises;
  }
}
</script>
</body>
</html>
